refactor(cli): fold the two backend query subcommands into one
